+++ etc/blocks/ui-examples/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @milaboratories/milaboratories.ui-examples.ui
 
+## 1.11.14
+
+### Patch Changes
+
+- Updated dependencies [b4f9e33]
+  - @platforma-sdk/model@1.76.5
+  - @milaboratories/milaboratories.ui-examples.model@1.6.14
+
 ## 1.11.13
 
 ### Patch Changes