+++ tools/pl-bootstrap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @platforma-sdk/bootstrap
 
+## 5.3.25
+
+### Patch Changes
+
+- Updated dependencies [33b2425]
+  - @milaboratories/pl-deployments@2.17.18
+
 ## 5.3.24
 
 ### Patch Changes